requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jiter==0.8.2
openai==1.59.8
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pyOpenSSL==25.0.0
requests==2.32.3
service-identity==24.2.0
setuptools==75.8.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.1
zope.interface==7.2
websocket-client~=1.8.0